workspace/test_simple_game/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]